Changes in e2e tests: - Create TestBase class and exctract common logic there - Simplify all e2e test suites
